+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Introduced the `mv` command, which renames a note file and also updates any links that reference the old file name in the front-matter of other notes.
+
 ## Features
 
 - __[Breaking Change]__ The `tag` and `untag` command now support tagging/untagging of multiple files - e.g. `tag <your tag> <file1> <file2> ...` is now possible. This is a breaking change because the order of the tag and file arguments on these commands has swapped.